--- backend/ddang/src/main/java/com/ddang/ddang/chat/presentation/dto/CreateMessageResponse.java
@@ -1,4 +1,0 @@
-package com.ddang.ddang.chat.presentation.dto;
-
-public record CreateMessageResponse(Long id) {
-}
