+++ backend/ddang/src/main/java/com/ddang/ddang/chat/presentation/dto/request/CreateMessageRequest.java
@@ -1,4 +1,4 @@
-package com.ddang.ddang.chat.presentation.dto;
+package com.ddang.ddang.chat.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
