+++ backend/ddang/src/main/java/com/ddang/ddang/chat/presentation/dto/response/CreateMessageResponse.java
@@ -1,0 +1,4 @@
+package com.ddang.ddang.chat.presentation.dto.response;
+
+public record CreateMessageResponse(Long id) {
+}
